rename: screens 폴더 이름 공백문자 제거
