+++ gymi/lib/screens/exercise/exercise_level1_screen.dart
@@ -1,23 +1,24 @@
 import 'dart:async';
-import 'package:eyedid_flutter_example/%08screens/exercise/exercise_intro.dart';
-import 'package:eyedid_flutter_example/%08screens/exercise/exercise_level4_screen.dart';
+import 'package:eyedid_flutter_example/screens/exercise/exercise_intro.dart';
+import 'package:eyedid_flutter_example/screens/exercise/exercise_level2_screen.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:google_fonts/google_fonts.dart';
 import '../../service/gaze_tracker_service.dart';
 import 'package:audioplayers/audioplayers.dart';
 
-class ExerciseLevel3Stage extends StatefulWidget {
+class ExerciseLevel1Stage extends StatefulWidget {
   final bool isVibrant;
   final bool isSingleMode;
-  const ExerciseLevel3Stage(
+
+  const ExerciseLevel1Stage(
       {super.key, this.isVibrant = true, this.isSingleMode = false});
 
   @override
-  State<ExerciseLevel3Stage> createState() => _ExerciseLevel3StageState();
+  State<ExerciseLevel1Stage> createState() => _ExerciseLevel1StageState();
 }
 
-class _ExerciseLevel3StageState extends State<ExerciseLevel3Stage>
+class _ExerciseLevel1StageState extends State<ExerciseLevel1Stage>
     with WidgetsBindingObserver {
   final _gazeService = GazeTrackerService();
   final AudioPlayer _audioPlayer = AudioPlayer();
@@ -25,14 +26,14 @@ class _ExerciseLevel3StageState extends State<ExerciseLevel3Stage>
   double _x = 0.0;
   double _y = 0.0;
 
-  String _currentTarget = 'left-up'; // 'left-up' 또는 'right-down'
+  String _currentTarget = 'up'; // 'up' 또는 'down'
   DateTime? _gazeStartTime;
   Timer? _gazeTimer;
   Timer? _progressTimer;
 
   bool _showCompletionMessage = false;
   bool _screenActive = true;
-  final bool _showTrackingFocus = true; // 🔥 final 제거해서 토글 가능
+  final bool _showTrackingFocus = true;
 
   final int _dwellTime = 3000; // 3초
   final int _totalSessionTime = 30; // 30초
@@ -55,19 +56,16 @@ class _ExerciseLevel3StageState extends State<ExerciseLevel3Stage>
 
     _startTime = DateTime.now();
 
-    // 🔥 progressTimer만 사용 (sessionTimer 제거)
+    // ✅ progressTimer만 사용 (sessionTimer 삭제)
     _progressTimer = Timer.periodic(const Duration(milliseconds: 100), (timer) {
       final elapsed = DateTime.now().difference(_startTime).inMilliseconds;
-
       if (!mounted) {
         timer.cancel();
         return;
       }
-
       setState(() {
         _progress = (elapsed / (_totalSessionTime * 1000)).clamp(0.0, 1.0);
       });
-
       if (_progress >= 1.0) {
         timer.cancel();
         _showCompletion();
@@ -98,7 +96,7 @@ class _ExerciseLevel3StageState extends State<ExerciseLevel3Stage>
     _gazeTimer?.cancel();
     _progressTimer?.cancel();
     _audioPlayer.dispose();
-    _gazeService.setShowOverlay(true);
+    // _gazeService.setShowOverlay(true);
 
     WidgetsBinding.instance.removeObserver(this);
 
@@ -123,32 +121,19 @@ class _ExerciseLevel3StageState extends State<ExerciseLevel3Stage>
         _dotSize = data['size'];
       });
 
-      _detectDiagonal();
+      _detectUpDown();
     });
   }
 
-  void _detectDiagonal() {
+  void _detectUpDown() {
     if (_showCompletionMessage) return;
 
-    final screenWidth = MediaQuery.of(context).size.width;
     final screenHeight = MediaQuery.of(context).size.height;
-
-    final centerX = screenWidth / 2;
     final centerY = screenHeight / 2;
 
-    bool inTarget = false;
+    String detectedDirection = _y < centerY ? 'up' : 'down';
 
-    if (_currentTarget == 'left-up') {
-      if (_x < centerX && _y < centerY) {
-        inTarget = true;
-      }
-    } else {
-      if (_x > centerX && _y > centerY) {
-        inTarget = true;
-      }
-    }
-
-    if (inTarget) {
+    if (detectedDirection == _currentTarget) {
       if (_gazeStartTime == null) {
         _gazeStartTime = DateTime.now();
         _startGazeTimer();
@@ -167,7 +152,7 @@ class _ExerciseLevel3StageState extends State<ExerciseLevel3Stage>
       HapticFeedback.mediumImpact();
 
       setState(() {
-        _currentTarget = _currentTarget == 'left-up' ? 'right-down' : 'left-up';
+        _currentTarget = _currentTarget == 'up' ? 'down' : 'up';
         _gazeStartTime = null;
       });
     });
@@ -199,7 +184,7 @@ class _ExerciseLevel3StageState extends State<ExerciseLevel3Stage>
           Navigator.pushReplacement(
             context,
             MaterialPageRoute(
-              builder: (context) => ExerciseLevel4Intro(
+              builder: (context) => ExerciseLevel2Intro(
                   isVibrant: widget.isVibrant, isSingleMode: false),
             ),
           );
@@ -231,7 +216,7 @@ class _ExerciseLevel3StageState extends State<ExerciseLevel3Stage>
                   children: const [
                     TextSpan(
                         text:
-                            "Look bottom-right to top-left until you see the \ngreen check. (30s)"),
+                            "Move your eyes up and down until you see the green check.\nKeep your head still. (30 seconds)"),
                   ],
                 ),
               ),
@@ -260,17 +245,17 @@ class _ExerciseLevel3StageState extends State<ExerciseLevel3Stage>
                     mainAxisAlignment: MainAxisAlignment.center,
                     children: [
                       Icon(
-                        _currentTarget == 'left-up'
-                            ? Icons.north_west
-                            : Icons.south_east,
+                        _currentTarget == 'up'
+                            ? Icons.arrow_upward
+                            : Icons.arrow_downward,
                         color: Colors.white,
                         size: 200,
                       ),
                       const SizedBox(height: 20),
                       Text(
-                        _currentTarget == 'left-up'
-                            ? "Move your eyes bottom-right"
-                            : "Move your eyes top-left",
+                        _currentTarget == 'up'
+                            ? "Move your eyes up "
+                            : "Move your eyes down",
                         style: GoogleFonts.roboto(
                             color: Colors.white,
                             fontSize: 40,
@@ -303,7 +288,7 @@ class _ExerciseLevel3StageState extends State<ExerciseLevel3Stage>
             ),
           ),
 
-          // 🔥 오른쪽 상단: 진행 인디케이터
+          // 🔥 오른쪽 상단: 진행도 인디케이터
           Positioned(
             top: 40,
             right: 40,
@@ -330,7 +315,7 @@ class _ExerciseLevel3StageState extends State<ExerciseLevel3Stage>
             ),
           ) /*,
 
-          // 🔥 왼쪽 상단 두번째 줄 : Focus 버튼
+          // 🔥 왼쪽 상단 두 번째 줄: Focus 버튼
           Positioned(
             left: 40,
             top: 120,
@@ -377,7 +362,7 @@ class _ExerciseLevel3StageState extends State<ExerciseLevel3Stage>
             left: 0,
             right: 0,
             child: Text(
-              "Level 3",
+              "Level 1",
               textAlign: TextAlign.center,
               style: GoogleFonts.roboto(
                   color: Colors.white,
@@ -392,17 +377,17 @@ class _ExerciseLevel3StageState extends State<ExerciseLevel3Stage>
   }
 }
 
-class ExerciseLevel3Intro extends StatefulWidget {
+class ExerciseLevel1Intro extends StatefulWidget {
   final bool isVibrant;
   final bool isSingleMode;
-  const ExerciseLevel3Intro(
+  const ExerciseLevel1Intro(
       {super.key, this.isVibrant = true, this.isSingleMode = false});
 
   @override
-  State<ExerciseLevel3Intro> createState() => _ExerciseLevel3IntroState();
+  State<ExerciseLevel1Intro> createState() => _ExerciseLevel1IntroState();
 }
 
-class _ExerciseLevel3IntroState extends State<ExerciseLevel3Intro> {
+class _ExerciseLevel1IntroState extends State<ExerciseLevel1Intro> {
   @override
   void initState() {
     super.initState();
@@ -413,10 +398,8 @@ class _ExerciseLevel3IntroState extends State<ExerciseLevel3Intro> {
         Navigator.pushReplacement(
           context,
           MaterialPageRoute(
-            builder: (context) => ExerciseLevel3Stage(
-              isVibrant: widget.isVibrant,
-              isSingleMode: widget.isSingleMode,
-            ),
+            builder: (context) => ExerciseLevel1Stage(
+                isVibrant: widget.isVibrant, isSingleMode: widget.isSingleMode),
           ),
         );
       }
@@ -444,7 +427,7 @@ class _ExerciseLevel3IntroState extends State<ExerciseLevel3Intro> {
           ),
           Center(
             child: Text(
-              "Level 3",
+              "Level 1",
               textAlign: TextAlign.center,
               style: GoogleFonts.roboto(
                   color: Colors.white,
