+++ gymi/lib/screens/exercise/exercise_level5_screen.dart
@@ -1,34 +1,34 @@
 import 'dart:async';
 import 'package:audioplayers/audioplayers.dart';
-import 'package:eyedid_flutter_example/%08screens/exercise/exercise_level9_screen.dart';
+import 'package:eyedid_flutter_example/screens/exercise/exercise_level6_screen.dart';
+import 'package:eyedid_flutter_example/screens/exercise/exercise_intro.dart';
 import 'package:eyedid_flutter_example/service/gaze_tracker_service.dart';
 import 'package:flutter/material.dart';
 import 'package:google_fonts/google_fonts.dart';
 import 'package:lottie/lottie.dart';
-import 'exercise_intro.dart'; // ← 다음 스크린으로 이동할 때 필요
 
-class ExerciseLevel8Stage extends StatefulWidget {
+class ExerciseLevel5Stage extends StatefulWidget {
   final bool isVibrant;
   final bool isSingleMode;
 
-  const ExerciseLevel8Stage({
+  const ExerciseLevel5Stage({
     super.key,
     this.isVibrant = true,
     this.isSingleMode = false,
   });
 
   @override
-  State<ExerciseLevel8Stage> createState() => _ExerciseLevel8StageState();
+  State<ExerciseLevel5Stage> createState() => _ExerciseLevel5StageState();
 }
 
-class _ExerciseLevel8StageState extends State<ExerciseLevel8Stage>
-    with SingleTickerProviderStateMixin {
+class _ExerciseLevel5StageState extends State<ExerciseLevel5Stage>
+    with SingleTickerProviderStateMixin, WidgetsBindingObserver {
   double _progress = 0.0;
   DateTime _startTime = DateTime.now();
   Timer? _progressTimer;
-  final AudioPlayer _audioPlayer = AudioPlayer();
   late final AnimationController _lottieController;
   bool _showCompletionMessage = false;
+  final AudioPlayer _audioPlayer = AudioPlayer();
 
   double _x = 0.0;
   double _y = 0.0;
@@ -37,38 +37,63 @@ class _ExerciseLevel8StageState extends State<ExerciseLevel8Stage>
 
   final _gazeService = GazeTrackerService();
   StreamSubscription<dynamic>? _gazeSubscription;
+
   @override
   void initState() {
     super.initState();
-    _gazeService.setShowOverlay(true);
-    _setupGazeTracking();
+    WidgetsBinding.instance.addObserver(this);
+
     _lottieController = AnimationController(vsync: this);
     _startTime = DateTime.now();
 
+    _gazeService.setShowOverlay(true);
+    _setupGazeTracking();
+
     _progressTimer = Timer.periodic(const Duration(milliseconds: 100), (timer) {
       final elapsed = DateTime.now().difference(_startTime).inMilliseconds;
-
       if (!mounted) {
         timer.cancel();
         return;
       }
-
       setState(() {
         _progress = (elapsed / (30 * 1000)).clamp(0.0, 1.0);
       });
-
       if (_progress >= 1.0) {
         timer.cancel();
-        _onSessionComplete();
+        _showCompletion();
       }
     });
   }
 
-  void _onSessionComplete() {
+  @override
+  void didChangeDependencies() {
+    super.didChangeDependencies();
+    WidgetsBinding.instance.addPostFrameCallback((_) {
+      _gazeService.updateContext(context); // 👈 overlay에 context 반영
+      _gazeService.refreshOverlay(); // 👈 overlay 다시 그리기
+    });
+  }
+
+  void _setupGazeTracking() {
+    _gazeSubscription = _gazeService.gazePositionStream.listen((data) {
+      if (!mounted) return;
+      setState(() {
+        _x = data['x'];
+        _y = data['y'];
+        _gazeColor = data['color'];
+        _dotSize = data['size'];
+      });
+    });
+  }
+
+  void _showCompletion() {
     setState(() {
       _showCompletionMessage = true;
     });
+
+    _gazeService.setShowOverlay(false); // 👈 다음 스테이지로 넘어가기 전에 끄기
     _audioPlayer.play(AssetSource('audio/correct.mp3'));
+
     Future.delayed(const Duration(seconds: 3), () {
       if (!mounted) return;
       if (widget.isSingleMode) {
@@ -83,7 +108,7 @@ class _ExerciseLevel8StageState extends State<ExerciseLevel8Stage>
         Navigator.pushReplacement(
           context,
           MaterialPageRoute(
-            builder: (context) => ExerciseLevel9Intro(
+            builder: (context) => ExerciseLevel6Intro(
                 isVibrant: widget.isVibrant, isSingleMode: false),
           ),
         );
@@ -92,34 +117,14 @@ class _ExerciseLevel8StageState extends State<ExerciseLevel8Stage>
   }
 
   @override
-  void didChangeDependencies() {
-    super.didChangeDependencies();
-    WidgetsBinding.instance.addPostFrameCallback((_) {
-      _gazeService.updateContext(context);
-      _gazeService.refreshOverlay();
-    });
-  }
-
-  @override
   void dispose() {
+    WidgetsBinding.instance.removeObserver(this);
     _gazeSubscription?.cancel();
     _lottieController.dispose();
     _progressTimer?.cancel();
     _audioPlayer.dispose();
-    _gazeService.setShowOverlay(true);
+    _gazeService.setShowOverlay(true); // 초기화용 true 설정
     super.dispose();
-  }
-
-  void _setupGazeTracking() {
-    _gazeSubscription = _gazeService.gazePositionStream.listen((data) {
-      if (!mounted) return;
-      setState(() {
-        _x = data['x'];
-        _y = data['y'];
-        _gazeColor = data['color'];
-        _dotSize = data['size'];
-      });
-    });
   }
 
   @override
@@ -145,7 +150,7 @@ class _ExerciseLevel8StageState extends State<ExerciseLevel8Stage>
                   children: const [
                     TextSpan(
                         text:
-                            "Follow the target point moving in an infinity shape. \nPlease keep your head still. (30s)"),
+                            "Draw a circle with your eyes clockwise until you see the green check. (30s)"),
                   ],
                 ),
               ),
@@ -169,25 +174,22 @@ class _ExerciseLevel8StageState extends State<ExerciseLevel8Stage>
                       ),
                     ],
                   )
-                : Transform.rotate(
-                    angle: 3.1416, // 180도 회전
-                    child: SizedBox(
-                      width: 700,
-                      height: 700,
-                      child: Lottie.asset(
-                        'assets/animations/infinity.json',
-                        controller: _lottieController,
-                        onLoaded: (composition) {
-                          _lottieController.duration =
-                              composition.duration * 2; // 속도 0.5배
-                          _lottieController.repeat();
-                        },
-                        repeat: true,
-                        fit: BoxFit.contain,
-                      ),
-                    ),
+                : Lottie.asset(
+                    'assets/animations/spin.json',
+                    controller: _lottieController,
+                    onLoaded: (composition) {
+                      _lottieController.duration = composition.duration * 2;
+                      _lottieController.repeat();
+                    },
+                    width: 700,
+                    height: 700,
+                    repeat: true,
+                    animate: true,
+                    fit: BoxFit.contain,
                   ),
           ),
+
+          // 상단 우측: 진행도
           Positioned(
             top: 40,
             right: 40,
@@ -204,29 +206,14 @@ class _ExerciseLevel8StageState extends State<ExerciseLevel8Stage>
                         const AlwaysStoppedAnimation<Color>(Colors.white),
                     strokeWidth: 6,
                   ),
-                  const Icon(
-                    Icons.access_time,
-                    color: Colors.black45,
-                    size: 32,
-                  ),
+                  const Icon(Icons.access_time,
+                      color: Colors.black45, size: 32),
                 ],
               ),
             ),
           ),
-          Positioned(
-            bottom: 50,
-            left: 0,
-            right: 0,
-            child: Text(
-              "Level 8",
-              textAlign: TextAlign.center,
-              style: GoogleFonts.roboto(
-                  color: Colors.white,
-                  fontSize: 36,
-                  fontStyle: FontStyle.normal,
-                  fontWeight: FontWeight.bold),
-            ),
-          ),
+
+          // 상단 좌측: 뒤로가기
           Positioned(
             top: 40,
             left: 40,
@@ -236,13 +223,28 @@ class _ExerciseLevel8StageState extends State<ExerciseLevel8Stage>
                 borderRadius: BorderRadius.circular(10),
               ),
               child: IconButton(
-                icon: const Icon(
-                  Icons.arrow_back,
-                  size: 30,
-                  color: Colors.black54,
-                ),
-                onPressed: () => Navigator.pop(context),
+                icon: const Icon(Icons.arrow_back,
+                    size: 30, color: Colors.black54),
+                onPressed: () {
+                  Navigator.pop(context);
+                },
               ),
+            ),
+          ),
+
+          // 하단 중앙: Level 표시
+          Positioned(
+            bottom: 50,
+            left: 0,
+            right: 0,
+            child: Text(
+              "Level 5",
+              textAlign: TextAlign.center,
+              style: GoogleFonts.roboto(
+                  color: Colors.white,
+                  fontSize: 36,
+                  fontStyle: FontStyle.normal,
+                  fontWeight: FontWeight.bold),
             ),
           ),
         ],
@@ -251,30 +253,29 @@ class _ExerciseLevel8StageState extends State<ExerciseLevel8Stage>
   }
 }
 
-class ExerciseLevel8Intro extends StatefulWidget {
+class ExerciseLevel5Intro extends StatefulWidget {
   final bool isVibrant;
   final bool isSingleMode;
-  const ExerciseLevel8Intro(
+  const ExerciseLevel5Intro(
       {super.key, this.isVibrant = true, this.isSingleMode = false});
 
   @override
-  State<ExerciseLevel8Intro> createState() => _ExerciseLevel8IntroState();
+  State<ExerciseLevel5Intro> createState() => _ExerciseLevel5IntroState();
 }
 
-class _ExerciseLevel8IntroState extends State<ExerciseLevel8Intro> {
+class _ExerciseLevel5IntroState extends State<ExerciseLevel5Intro> {
   @override
   void initState() {
     super.initState();
 
-    // 3초 후 자동 이동
     Future.delayed(const Duration(seconds: 3), () {
       if (mounted) {
         Navigator.pushReplacement(
           context,
           MaterialPageRoute(
-            builder: (context) => ExerciseLevel8Stage(
+            builder: (context) => ExerciseLevel5Stage(
               isVibrant: widget.isVibrant,
-              isSingleMode: widget.isSingleMode, // Single Mode 여부 전달
+              isSingleMode: widget.isSingleMode,
             ),
           ),
         );
@@ -289,13 +290,12 @@ class _ExerciseLevel8IntroState extends State<ExerciseLevel8Intro> {
           widget.isVibrant ? const Color(0xFFAEC7DF) : const Color(0xFFA38D7D),
       body: Stack(
         children: [
-          /// 1️⃣ PNG 이미지를 오른쪽 정렬(Align.end) + 자연스럽게 잘리도록 ClipRect 사용
           Align(
-            alignment: Alignment.centerRight, // 👉 오른쪽 정렬
+            alignment: Alignment.centerRight,
             child: ClipRect(
               child: Image.asset(
-                'assets/images/gymiBackground.png', // ✅ PNG 이미지 경로
-                fit: BoxFit.fitHeight, // ✅ 화면 크기에 맞춰 자연스럽게 채우기
+                'assets/images/gymiBackground.png',
+                fit: BoxFit.fitHeight,
                 width: 650,
                 height: 900,
               ),
@@ -303,7 +303,7 @@ class _ExerciseLevel8IntroState extends State<ExerciseLevel8Intro> {
           ),
           Center(
             child: Text(
-              "Level 8",
+              "Level 5",
               textAlign: TextAlign.center,
               style: GoogleFonts.roboto(
                   color: Colors.white,
