+++ gymi/lib/screens/eye_judge_screen.dart
@@ -3,8 +3,8 @@ import 'dart:io';
 import 'dart:typed_data';
 import 'dart:ui' as ui;
 import 'package:camera/camera.dart';
-import 'package:eyedid_flutter_example/%08screens/bad_screen.dart';
-import 'package:eyedid_flutter_example/%08screens/good_screen.dart';
+import 'package:eyedid_flutter_example/screens/bad_screen.dart';
+import 'package:eyedid_flutter_example/screens/good_screen.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/rendering.dart';
 import 'package:google_fonts/google_fonts.dart';
