+++ gymi/lib/screens/home_screen.dart
@@ -1,12 +1,12 @@
 import 'dart:async';
-import 'package:eyedid_flutter_example/%08screens/exercise/exercise_intro.dart';
+import 'package:eyedid_flutter_example/screens/exercise/exercise_intro.dart';
 
-import 'package:eyedid_flutter_example/%08screens/tutorial/tutorial.dart';
-import 'package:eyedid_flutter_example/%08screens/calibration_screen.dart';
-import 'package:eyedid_flutter_example/%08screens/exercise/exercise_level1_screen.dart';
-import 'package:eyedid_flutter_example/%08screens/game/game_screen.dart';
+import 'package:eyedid_flutter_example/screens/tutorial/tutorial.dart';
+import 'package:eyedid_flutter_example/screens/calibration_screen.dart';
+import 'package:eyedid_flutter_example/screens/exercise/exercise_level1_screen.dart';
+import 'package:eyedid_flutter_example/screens/game/game_screen.dart';
 import 'package:eyedid_flutter_example/service/gaze_tracker_service.dart';
-import 'package:eyedid_flutter_example/%08screens/tutorial/tutorial_intro.dart';
+import 'package:eyedid_flutter_example/screens/tutorial/tutorial_intro.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:google_fonts/google_fonts.dart';
