+++ gymi/lib/screens/tutorial/tutorial_ready_screen.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
-import 'package:eyedid_flutter_example/%08screens/tutorial/tutorial.dart';
-import 'package:eyedid_flutter_example/%08screens/setting_screen.dart';
+import 'package:eyedid_flutter_example/screens/tutorial/tutorial.dart';
+import 'package:eyedid_flutter_example/screens/setting_screen.dart';
 import 'package:google_fonts/google_fonts.dart';
 
 class TutorialReadyScreen extends StatefulWidget {
